+++ lazarus/tasks/transforms.py
@@ -13,15 +13,28 @@ class Transform(Task):
         return None
 
 
-class ExtractPostID(Transform):
+class CommentFilter(Transform):
+    def __init__(self, columns: List[str], *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.columns = set(columns)
+
     def __call__(self, msg):
+        try:
+            float(msg["sentiment"])
+        except:  # pylint: disable=bare-except
+            logger.error("Bad sentiment %s", msg["sentiment"])
+            return None
+
         msg_id = re.match(
             r"https://old.reddit.com/r/me_?irl/comments/([^/]+)/.*", msg["permalink"]
         ).groups()
         if msg_id is None:
             logger.error("Bad permalink %s", msg["permalink"])
             return
+
         msg["id"] = msg_id[0]
+        msg = {k: v for k, v in msg.items() if k in self.columns}
+
         return msg
 
 
@@ -41,6 +54,9 @@ class PostsMeanSentiment(Transform):
             for k, v in self.sentiment_scores.items()
         ]
 
+    def reset(self):
+        self.sentiment_scores = defaultdict(lambda: {"count": 0, "sum": 0})
+
 
 class PostsMeanScore(Transform):
     def __init__(self):
@@ -54,6 +70,10 @@ class PostsMeanScore(Transform):
 
     def collect(self):
         return [{"posts_mean_score": self.total_posts_scores / self.n_posts_scores}]
+
+    def reset(self):
+        self.n_posts_scores = 0
+        self.total_posts_scores = 0
 
 
 class FilterColumn(Transform):