+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-07-05
+### Added
+- Generic Node implementation to handle joins
+- Sentiment joiner
+- Comment filter
+- Post mean sentiment
+- Best meme downloader
+- Student comment filter
+- School Joiner
+
 ## [2.1.0] - 2022-07-05
 ### Added
 - Initial bully leader election algorithm